102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 
 void fibonacci(void)
 {
-	int a, b = 1, c = 1, i;
+	long long int a, b = 1, c = 1, i;
 
 	for (i = 0; i < 50; i++)
 	{
